requirements.txt
json
requests 
python-dotenv